+++ src/config/auth/auth.ts
@@ -2,7 +2,7 @@ import dotenv from 'dotenv';
 
 dotenv.config();
 
-const SERVER_TOKEN_EXPIRE = process.env.SERVER_PORT || 3600;
+const SERVER_TOKEN_EXPIRE = process.env.SERVER_PORT || 3000;
 const SERVER_TOKEN_SECRET = process.env.SERVER_PORT || 'coolIssuer';
 const SERVER_TOKEN_ISSUER = process.env.SERVER_PORT || 'superencryptedsecret';